+++ api/src/main/java/com/codeforcommunity/dto/school/SchoolContact.java
@@ -2,13 +2,13 @@ package com.codeforcommunity.dto.school;
 
 public class SchoolContact {
 
-  private Long id;
+  private Integer id;
   private String name;
   private String email;
   private String address;
   private String phone;
 
-  public SchoolContact(Long id, String name, String email, String address, String phone) {
+  public SchoolContact(Integer id, String name, String email, String address, String phone) {
     this.id = id;
     this.name = name;
     this.email = email;
@@ -16,11 +16,11 @@ public class SchoolContact {
     this.phone = phone;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
